backend/pyproject.toml
[project]
name = "weaver-backend"
version = "0.1.0"
description = "Backend for Weaver tool-card-first multi-agent platform"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.35.0",
  "pydantic>=2.9.0",
  "pydantic-settings>=2.5.0",
  "sqlalchemy>=2.0.35",
  "aiosqlite>=0.20.0",
  "psycopg[binary]>=3.2.3",
  "asyncpg>=0.30.0",
  "alembic>=1.13.0",
  "redis>=5.0.8",
  "arq>=0.26.1",
  "authlib>=1.3.2",
  "httpx>=0.27.2",
  "structlog>=24.4.0",
  "opentelemetry-api>=1.27.0",
  "opentelemetry-sdk>=1.27.0",
  "python-jose[cryptography]>=3.3.0",
  "google-api-python-client>=2.150.0",
  "google-auth>=2.35.0",
  "google-auth-oauthlib>=1.2.1",
  "aiohttp>=3.10.10",
  "langgraph>=0.2.35",
  "langchain-core>=0.3.10",
]

[project.scripts]
weaver-cli = "app.cli:main"

[dependency-groups]
dev = [
  "pytest>=8.3.3",
  "pytest-asyncio>=0.24.0",
  "httpx>=0.27.2",
  "ruff>=0.6.8",
  "mypy>=1.11.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
asyncio_mode = "auto"
